# swagger.json
{ 
   "swagger":"2.0",
   "info":{ 
      "version":"V1",
      "title":"Book API"
   },
   "paths":{ 
      "/v1/books/search":{ 
         "get":{ 
            "tags":[ 
               "Books"
            ],
            "summary":"Ricerca libri con testo libero",
            "operationId":"SearchBooksAsync",
            "consumes":[ 

            ],
            "produces":[ 
               "text/plain",
               "application/json",
               "text/json"
            ],
            "parameters":[ 
               { 
                  "name":"searchText",
                  "in":"query",
                  "description":"I parametri di ricerca",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"Success",
                  "schema":{ 
                     "uniqueItems":false,
                     "type":"array",
                     "items":{ 
                        "$ref":"#/definitions/BookListItem"
                     }
                  }
               }
            }
         }
      },
      "/v1/books/{key}":{ 
         "get":{ 
            "tags":[ 
               "Books"
            ],
            "summary":"I dettagli di un libro",
            "operationId":"GetBookDetailsAsync",
            "consumes":[ 

            ],
            "produces":[ 
               "text/plain",
               "application/json",
               "text/json"
            ],
            "parameters":[ 
               { 
                  "name":"key",
                  "in":"path",
                  "description":"La chiave del libro",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"Success",
                  "schema":{ 
                     "uniqueItems":false,
                     "type":"array",
                     "items":{ 
                        "$ref":"#/definitions/BookListItem"
                     }
                  }
               }
            }
         },
         "post":{ 
            "tags":[ 
               "Books"
            ],
            "summary":"Aggiunge una recensione a un libro",
            "operationId":"PostReviewAsync",
            "consumes":[ 
               "application/json-patch+json",
               "application/json",
               "text/json",
               "application/*+json"
            ],
            "produces":[ 
               "text/plain",
               "application/json",
               "text/json"
            ],
            "parameters":[ 
               { 
                  "name":"key",
                  "in":"path",
                  "description":"La chiave del libro",
                  "required":true,
                  "type":"string"
               },
               { 
                  "name":"bookReviewRequest",
                  "in":"body",
                  "description":"I paraemtri della richiesta",
                  "required":false,
                  "schema":{ 
                     "$ref":"#/definitions/BookReviewRequest"
                  }
               }
            ],
            "responses":{ 
               "201":{ 
                  "description":"Success",
                  "schema":{ 
                     "uniqueItems":false,
                     "type":"array",
                     "items":{ 
                        "$ref":"#/definitions/BookListItem"
                     }
                  }
               }
            }
         }
      }
   },
   "definitions":{ 
      "BookListItem":{ 
         "type":"object",
         "properties":{ 
            "title":{ 
               "type":"string"
            },
            "author":{ 
               "type":"string"
            },
            "key":{ 
               "type":"string"
            },
            "shortDescription":{ 
               "type":"string"
            }
         }
      },
      "BookReviewRequest":{ 
         "required":[ 
            "reviewText",
            "bookRating"
         ],
         "type":"object",
         "properties":{ 
            "reviewText":{ 
               "description":"Il testo della recensione",
               "type":"string"
            },
            "bookRating":{ 
               "description":"Il rating della recensione (da 1 a 5)",
               "enum":[ 
                  "OneStar",
                  "TwoStars",
                  "ThreeStars",
                  "FourStars",
                  "FiveStars"
               ],
               "type":"string"
            }
         }
      }
   }
}